Priority, Progress and Recurring
Added task priority, tasks are now ordered based on date and priority. Category progress based on the amount of tasks completed vs. not completed in the category. Recurring task that creates a new one every time the old task is complete.

Bug Notice: When marking a recurring task as complete, the original task quickly marks to "Todo" and a new task is made, but you can't remove the old task and it won't stay complete, so you can unintentionally continue creating duplicate tasks from the first task.